requirements.txt
Pillow>=10.0.0
pathlib2>=2.3.7 